Fixes disappearance of P25 and %>=Q30 from summary, qc/lineage creation fail on NA consensus seq name. Revises folder structure for unit and regression tests, expands regression tests

+++ tests/test_seq_run_acceptance.py
@@ -6,17 +6,23 @@ from qc.seq_run_acceptance import make_collected_acceptances_tsv
 
 class SeqRunAcceptanceTest(FileTestCase):
     def test_merge_tables(self):
-        expected_results_fp = f"{self.test_qc_dir}/" \
+        expected_results_fp = f"{self.gs_qc_dir}/" \
                               f"2021-02-08-ARTIC-acceptance.tsv"
-        output_fp = f"{self.test_temp_qc_dir}/temp_seq_run_acceptance.csv"
-        arg_list = ["seq_run_acceptance.py", self.test_samples_dir, output_fp]
+        output_fp = f"{self.test_temp_dir}/temp_seq_run_acceptance.tsv"
+        arg_list = ["seq_run_acceptance.py", self.gs_samples_dir, output_fp]
 
+        out_is_file = out_matches = False
         try:
             make_collected_acceptances_tsv(arg_list)
-            self.assertTrue(os.path.isfile(output_fp))
-            self.assertTrue(filecmp.cmp(output_fp, expected_results_fp))
+
+            out_is_file = os.path.isfile(output_fp)
+            self.assertTrue(out_is_file)
+
+            out_matches = filecmp.cmp(output_fp, expected_results_fp)
+            self.assertTrue(out_matches)
         finally:
-            try:
-                os.remove(output_fp)
-            except OSError:
-                pass
+            if out_is_file and out_matches:
+                try:
+                    os.remove(output_fp)
+                except OSError:
+                    pass
